requirements/static-analysis-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile --output-file=requirements/static-analysis-requirements.txt --resolver=backtracking requirements/static-analysis.in
#
alabaster==0.7.13
    # via sphinx
asgiref==3.6.0
    # via django
astroid==2.13.2
    # via
    #   pylint
    #   pylint-celery
    #   pylint-flask
    #   requirements-detector
attrs==22.2.0
    # via pytest
babel==2.12.1
    # via sphinx
backports-zoneinfo==0.2.1
    # via django
bandit==1.7.4
    # via prospector
black==22.12.0
    # via -r requirements/static-analysis.in
build==0.10.0
    # via pyroma
certifi==2022.12.7
    # via requests
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via black
coverage[toml]==7.0.5
    # via pytest-cov
dill==0.3.6
    # via pylint
django==4.1.7
    # via
    #   -r requirements/static-analysis.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements/static-analysis.in
docutils==0.19
    # via
    #   pyroma
    #   sphinx
dodgy==0.2.1
    # via prospector
exam==0.10.6
    # via -r requirements/static-analysis.in
exceptiongroup==1.1.0
    # via pytest
flake8==5.0.4
    # via
    #   flake8-polyfill
    #   prospector
flake8-polyfill==1.0.2
    # via pep8-naming
gitdb==4.0.10
    # via gitpython
gitpython==3.1.30
    # via
    #   bandit
    #   prospector
idna==3.4
    # via requests
imagesize==1.4.1
    # via sphinx
importlib-metadata==6.0.0
    # via sphinx
iniconfig==2.0.0
    # via pytest
isort==5.11.4
    # via pylint
jinja2==3.1.2
    # via sphinx
lazy-object-proxy==1.9.0
    # via astroid
markupsafe==2.1.2
    # via jinja2
mccabe==0.7.0
    # via
    #   flake8
    #   prospector
    #   pylint
mock==5.0.1
    # via exam
mypy==0.991
    # via prospector
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
packaging==21.3
    # via
    #   build
    #   prospector
    #   pytest
    #   requirements-detector
    #   sphinx
pathspec==0.11.0
    # via black
pbr==5.11.1
    # via stevedore
pep517==0.13.0
    # via pyroma
pep8-naming==0.10.0
    # via prospector
platformdirs==2.6.2
    # via
    #   black
    #   pylint
pluggy==1.0.0
    # via pytest
poetry-semver==0.1.0
    # via requirements-detector
prospector[with_everything]==1.8.4
    # via -r requirements/static-analysis.in
psycopg2-binary==2.9.5
    # via -r requirements/static-analysis.in
pycodestyle==2.9.1
    # via
    #   flake8
    #   prospector
pydocstyle==6.2.3
    # via prospector
pyflakes==2.5.0
    # via
    #   flake8
    #   prospector
pygments==2.14.0
    # via
    #   pyroma
    #   sphinx
pylint==2.15.10
    # via
    #   -r requirements/static-analysis.in
    #   prospector
    #   pylint-celery
    #   pylint-django
    #   pylint-flask
    #   pylint-plugin-utils
pylint-celery==0.3
    # via prospector
pylint-django==2.5.3
    # via prospector
pylint-flask==0.6
    # via prospector
pylint-plugin-utils==0.7
    # via
    #   prospector
    #   pylint-celery
    #   pylint-django
    #   pylint-flask
pyparsing==3.0.9
    # via packaging
pyproject-hooks==1.0.0
    # via build
pyroma==4.1
    # via prospector
pytest==7.2.0
    # via
    #   -r requirements/static-analysis.in
    #   pytest-cov
    #   pytest-django
pytest-cov==4.0.0
    # via -r requirements/static-analysis.in
pytest-django==4.5.2
    # via -r requirements/static-analysis.in
pytz==2022.7.1
    # via
    #   babel
    #   djangorestframework
pyyaml==6.0
    # via
    #   bandit
    #   prospector
requests==2.28.2
    # via
    #   pyroma
    #   sphinx
requirements-detector==1.0.3
    # via prospector
setoptconf-tmp==0.3.1
    # via prospector
smmap==5.0.0
    # via gitdb
snowballstemmer==2.2.0
    # via
    #   pydocstyle
    #   sphinx
sphinx==6.1.3
    # via
    #   -r requirements/static-analysis.in
    #   sphinx-rtd-theme
    #   sphinxnotes-strike
sphinx-rtd-theme==0.5.1
    # via -r requirements/static-analysis.in
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sphinxnotes-strike==1.2
    # via -r requirements/static-analysis.in
sqlparse==0.4.3
    # via django
stevedore==4.1.1
    # via bandit
toml==0.10.2
    # via
    #   prospector
    #   requirements-detector
    #   vulture
tomli==2.0.1
    # via
    #   black
    #   build
    #   coverage
    #   mypy
    #   pep517
    #   pylint
    #   pyproject-hooks
    #   pytest
tomlkit==0.11.6
    # via pylint
trove-classifiers==2023.1.12
    # via pyroma
typing-extensions==4.4.0
    # via
    #   astroid
    #   black
    #   mypy
    #   pylint
urllib3==1.26.14
    # via requests
vulture==2.7
    # via prospector
wheel==0.38.4
    # via pyroma
wrapt==1.14.1
    # via astroid
zipp==3.15.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
